+++ web/src/app/release-controls.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, Optional, Inject,Output,EventEmitter } from '@angular
 import { Router }            from '@angular/router';
 import { MdDialog, MdDialogRef, MD_DIALOG_DATA } from '@angular/material';
 import {NotificationsService} from './angular2-notifications/simple-notifications.module';
+import { AceDialogComponent } from './ace-dialog.component';
 
 import { Observable }        from 'rxjs/Observable';
 import { Subject }           from 'rxjs/Subject';
@@ -54,7 +55,7 @@ export class ReleaseControlsComponent {
   openEditDialog(rel: Release) {
     console.log(rel);
     let configData = rel.config.raw ? rel.config.raw.trim():"";
-    const dialogRef = this._dialog.open(DialogContentComponent, {
+    const dialogRef = this._dialog.open(AceDialogComponent, {
       data: {'config':configData, 'values':rel.chart.values.raw},
     });
     dialogRef.afterClosed().subscribe(result => {
@@ -78,108 +79,4 @@ export class ReleaseControlsComponent {
   toggleLoad(): void {
     this.loading = this.loading ? false: true; 
   }
-}
-
-
-
-@Component({
-  template: `
-    <ace-editor
-      #editorInput
-      [(text)]="data.config"
-      [mode]="'yaml'"></ace-editor>
-    <pre>{{ data.values }}</pre>
-    <br />
-    <button md-button (click)="dialogRef.close(editorInput.value)">
-      <md-icon>save</md-icon> save
-    </button>
-    <button md-button (click)="dialogRef.close()">
-      <md-icon>cancel</md-icon> cancel
-    </button>
-  `,
-  styles: [`
-    ace-editor {
-      height: 10em;
-      width: auto;
-      font-family: "Courier New", Courier, monospace !important;
-    }
-    pre {
-      width: auto;
-      height: 10em;
-      overflow: auto;
-      background-color: #eeeeee;
-      word-break: normal !important;
-      word-wrap: normal !important;
-      white-space: pre !important;
-    }
-  `],
-})
-export class DialogContentComponent {
-  //text: string = 
-  constructor( 
-    @Optional() public dialogRef: MdDialogRef<DialogContentComponent>,
-    @Inject(MD_DIALOG_DATA) public data: any
-  ) {}
-}
-
-import { Pipe, PipeTransform } from '@angular/core';
-
-@Pipe({name: 'unepoch'})
-export class UnEpochPipe implements PipeTransform {
-  transform(value: number, args: string[]): any {
-    if (!value) return value;
-    var t = new Date(1970, 0, 1); // Epoch
-    t.setSeconds(value);
-
-    return t.toLocaleDateString("en-US");
-  }
-}
-
-@Pipe({name: 'status'})
-export class StatusStringPipe implements PipeTransform {
-  transform(value: number, args: string[]): any {
-    let statuses = [
-      "UNKNOWN",
-      "DEPLOYED",
-      "DELETED",
-      "SUPERSEDED",
-      "FAILED",
-      "DELETING"
-    ];
-    
-    if (!value) return value;
-    if (value > statuses.length) return "UNKNOWN";
-    
-    return statuses[value];
-  }
-}
-
-import {DomSanitizer, SafeHtml, SafeStyle, SafeScript, SafeUrl, SafeResourceUrl} from '@angular/platform-browser';
-
-@Pipe({
-	name: 'safe'
-})
-export class SafePipe {
-
-	constructor(protected _sanitizer: DomSanitizer) {
-
-	}
-
-	public transform(value: string, type: string): SafeHtml | SafeStyle | SafeScript | SafeUrl | SafeResourceUrl {
-		switch (type) {
-			case 'html':
-				return this._sanitizer.bypassSecurityTrustHtml(value);
-			case 'style':
-				return this._sanitizer.bypassSecurityTrustStyle(value);
-			case 'script':
-				return this._sanitizer.bypassSecurityTrustScript(value);
-			case 'url':
-				return this._sanitizer.bypassSecurityTrustUrl(value);
-			case 'resourceUrl':
-				return this._sanitizer.bypassSecurityTrustResourceUrl(value);
-			default:
-				throw new Error(`Unable to bypass security for invalid type: ${type}`);
-		}
-	}
-
 }
